+++ example/integration_test/constants.dart
@@ -3,4 +3,7 @@ final TEST_CROSS_PLATFORM_URL_1 = Uri.parse('https://flutter.dev');
 final TEST_CROSS_PLATFORM_URL_2 = Uri.parse('https://www.bing.com/');
 final TEST_URL_1 = Uri.parse('https://github.com/flutter');
 final TEST_URL_2 = Uri.parse('https://www.google.com/');
+final TEST_URL_3 = Uri.parse('https://github.com/pichillilorenzo/flutter_inappwebview');
+final TEST_URL_4 = Uri.parse('https://www.youtube.com/');
 final TEST_WEB_PLATFORM_URL_1 = Uri.parse(Uri.base.toString().replaceFirst("/#/", "/") + 'page.html');
+final TEST_NOT_A_WEBSITE_URL = Uri.parse('https://www.notawebsite..com');