+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Added support for `onPermissionRequest` event on iOS 15.0+
 - Updated `getMetaThemeColor` on iOS 15.0+
 
+### BREAKING CHANGES
+
+- On Android, the `InAppWebView` widget uses hybrid composition by default (`useHybridComposition: true`).
+
 ## 5.4.3
 
 - Added Bitwise OR operator support for `AndroidActionModeMenuItem` class